pyproject.toml
[tool.poetry]
name = "tomsk-special-economic-zone-bot"
version = "0.1.0"
description = "Чат-бот Особой экономической зоны технико-внедренческого типа \"Томск\""
authors = ["Victor <MarkerViktor@outlook.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
dependency-injector = "^4.41.0"
databases = {extras = ["asyncpg"], version = "^0.8.0"}
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
mypy = "^1.6.0"
ruff = "^0.1.0"
asyncpg-stubs = "^0.28.0"

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120
target-version = ['py311']
preview = true
